Updated to use json4s jackson backend.
The json4s-native backend is broken in several ways. It encodes all JSON, which is allowed, but non-standard since JSON is required to be UTF-8 encoded, and it fails to properly escape UTF-8 control characters, which can often show up in localized strings. See: https://github.com/json4s/json4s/issues/333#issuecomment-179934084

The breakage introduced in json4s 3.3.x was rolled back and corrected in PR339 as part of 3.4.0-SNAPSHOT, but that hasn't been published yet. Using the Jackson backend works around this because the breakage is limited to the native implementation.